(refactor) Create query component directory
* Anticipating robust implementation
